factor out graphviz from web_c
